[doc] Modularize Redux documents
