+++ package.json
@@ -10,19 +10,19 @@
     "supertest": "^2.0.0"
   },
   "dependencies": {
+    "axios": "^0.13.1",
     "babel-polyfill": "^6.9.1",
     "body-parser": "^1.15.2",
     "csv": "^1.1.0",
     "es6-promise": "^3.2.1",
     "express": "^4.14.0",
+    "express-csv-middleware": "^0.1.0",
     "isomorphic-fetch": "^2.2.1",
     "react": "^15.2.1",
     "react-dom": "^15.2.1",
     "react-redux": "^4.4.5",
     "redux": "^3.5.2",
     "redux-logger": "^2.6.1",
-    "redux-thunk": "^2.1.0",
-    "express-csv-middleware": "^0.1.0",
     "redux-promise": "^0.5.3",
     "redux-thunk": "^2.1.0"
   },
