+++ src/actions/categoryActions.js
@@ -1,0 +1,4 @@
+import Axios from 'axios'
+
+export const ADD_CATEGORY = 'ADD_CATEGORY';
+export const CATEGORY_UPDATE = 'CATEGORY_UPDATE';
